+++ modules/VaccineRegistration/Http/Controllers/ArticleController.php
@@ -52,6 +52,7 @@ class ArticleController extends Controller
             'Chăm Sóc Trẻ Em',
             'Tiêm Chủng Người Lớn',
             'Tiêm Phòng Mẹ Bầu',
+            'Góc Chuyên Gia',
         ]);
 
         return view('vaccine::articles.index', compact('articles', 'categories', 'hotNews'));